Stream segment neighbours

For each segment, report those segments that are connected to one or multiple input segments within a specified neighbour order, with the option to summarize attributes across these segments.

Usage

segment_neighbours(
  g,
  segmentID = NULL,
  variable = NULL,
  stat = NULL,
  attach_only = F,
  order = 5,
  mode = "in",
  n_cores = 1,
  maxsize = 1500
)

Arguments

g

A directed graph (igraph object).

segmentID

The input segment IDs as a numerical vector for which to search the connected segments.

variable

Optional. One or more attribute(s) or variable(s) of the input graph that should be reported for each output segmentID ("to_stream").

stat

One of mean, median, min, max, sd (without quotes). Aggregates (or summarizes) the variables for the neighbourhood of each input segment ("stream", e.g., the average land cover in the next five upstream segments or sub-catchments).

attach_only

Logical. If TRUE then the selected variables will be only attached to each for each segment without any further aggregation.

order

The neighbouring order. Order=1 would be immediate neighbours of the input segementID, order=2 would be the order 1 plus the immediate neighbours of the segementIDs in order 1.

mode

One of "in", "out", or "all". "in" reports only upstream neighbour segments, "out" reports only the downstream segments, and "all" does both.

n_cores

Optional. Specify the number of CPUs for internal parallelization in the case of multiple stream segments / outlets. Defaults to 1. In case the graph is very large, and many segments are used as an input, setting n_cores to a higher value can speed up the coputatation. This comes however at the cost of possible RAM limitations and even slower processing since the large data will be copied to each CPU.

maxsize

Optional. Specify the maximum size of the data passed to the parallel backend in MB. Defaults to 1500 (1.5 GB). Consider a higher value for large study areas (more than one 20°x20° tile).

Details

This function can also be used to create the connectivity table for Marxan by using agg_var="length" and attach_only=TRUE. The resulting table reports the connectivity from each segment, along with the stream length for all connected segments.