Crop raster to extent

This function crops an input raster layer directly on disk, i.e. the input layer does not need to be loaded into R. The raster .tif is cropped to a polygon border line if a vector layer (cutline source) is provided, otherwise if a bounding box is provided (xmin, ymin, xmax, ymax coordinates or a spatial object from which to extract a bounding box), the raster is cropped to the extent of the bounding box. At least a cutline source (vector_layer) or a bounding box (bounding_box) should be provided. The output is always written to disk, and can be optionally loaded into R as a SpatRaster (terra package) object (using read = TRUE).

Usage

crop_to_extent(
  raster_layer,
  vector_layer = NULL,
  bounding_box = NULL,
  out_dir,
  file_name,
  read = TRUE,
  quiet = TRUE
)

Arguments

raster_layer

character. Full path to the input raster .tif layer.

vector_layer

character. Full path to a vector layer that is used as a cutline data source (similar to a mask operation).

bounding_box

numeric vector of the coordinates of the corners of a bounding box (xmin, ymin, xmax, ymax), SpatRaster, SpatVector, or other spatial object.

out_dir

character. The directory where the output will be stored.

file_name

character. Name of the cropped output raster .tif file.

read

logical. If TRUE, the cropped raster .tif layer gets read into R. If FALSE, the layer is only stored on disk. Default is TRUE.

quiet

logical. If FALSE, the standard output will be printed. Default is TRUE.

Value

The function returns always a .tif raster file written to disk. Optionally, a SpatRaster (terra object) can be loaded into R with read = TRUE.

Author

Yusdiel Torres-Cambas

Examples

# Download test data into the temporary R folder
# or define a different directory
my_directory <- tempdir()
download_test_data(my_directory)

# Define full path to the input raster .tif layer ans vector layer
spi_raster <- paste0(my_directory, "/hydrography90m_test_data",
                     "/spi_1264942.tif")
basin_vector <- paste0(my_directory, "/hydrography90m_test_data",
                      "/basin_59.gpkg")

# Crop the Stream Power Index to the basin
spi_basin <- crop_to_extent(raster_layer = spi_raster,
                            vector_layer = basin_vector,
                            out_dir = my_directory,
                            file_name = "spi_basin_cropped.tif",
                            read = TRUE)