Download files of the Hydrography90m dataset

The function downloads files of the Hydrography90m dataset, available at https://public.igb-berlin.de/index.php/s/agciopgzXjWswF4?path=%2F . The files will be stored in the folder architecture of the above domain. Multiple regular tile or regional unit files can be requested in a single call of the function. The tile or regional unit IDs can be obtained using the functions "get_tile_id" and "get_regional_unit_id" respectively.

Usage

download_tiles(
  variable,
  file_format = "tif",
  tile_id = NULL,
  reg_unit_id = NULL,
  global = FALSE,
  download_dir = "."
)

Arguments

variable

character vector of variable names.

file_format

character. Format of the requested file ("tif" or "gpkg").

tile_id

character vector. The IDs of the requested tiles.

reg_unit_id

character vector. The IDs of the requested regional units.

global

logical. If TRUE, the global extent file is downloaded. Default is FALSE.

download_dir

character. The directory where the files will be downloaded. Default is the working directory.

Details

Variable type	Variable	Abbreviation	Description	File_format	Unit

Flow	Flow accumulation	flow	Flow accumulation	tif	km²

Flow	Flow accumulation (positive values)	flowpos	Flow accumulation with positive values	tif	km²

Stream slope	Cell maximum curvature	slope_curv_max_dw_cel	Cell maximum curvature (between highest upstream cell,focal cell and downstream cell)	tif	m⁻¹

Stream slope	Cell minimum curvature	slope_curv_min_dw_cel	Cell minimum curvature(between lowest upstream cell,focal cell and downstream cell)	tif	m⁻¹

Stream slope	Cell elevation difference	slope_elv_dw_cel	Cell elevation difference(between focal cell and downstream cell)	tif	m

Stream slope	Cell gradient	slope_grad_dw_cel	Cell gradient	tif

Stream distance	Shortest distance to drainage divide	stream_dist_up_near	Shortest upstream distance between focal grid cell and the nearest sub-catchment drainage divide	tif	m

Stream distance	Longest distance to drainage divide	stream_dist_up_farth	Longest upstream distance between focal grid cell and the nearest sub-catchment drainage divide	tif	m

Stream distance	Nearest down stream stream grid cell	stream_dist_dw_near	Distance between focal grid cell and its nearest down stream stream grid cell	tif	m

Stream distance	Outlet grid cell in the network	outlet_dist_dw_basin	Distance between focal grid cell and the outlet grid cell in the network	tif	m

Stream distance	Down stream stream node grid cell	outlet_dist_dw_scatch	Distance between focal grid cell and the down stream stream node grid cell	tif	m

Stream distance	Euclidean distance	stream_dist_proximity	Euclidean distance between focal grid cell and the stream network (in meters)	tif	m

Elevation difference	Shortest path	stream_diff_up_near	Elevation difference of the shortest path from focal grid cell to the sub-catchment drainage divide	tif	m

Elevation difference	Longest path	stream_diff_up_farth	Elevation difference of the longest path from focal grid cell to the sub-catchment drainage divide	tif	m

Elevation difference	Nearest downstream stream pixel	stream_diff_dw_near	Elevation difference between focal grid cell and its nearest downstream stream pixel	tif	m

Elevation difference	Outlet grid cell in the network	outlet_diff_dw_basin	Elevation difference between focal grid cell and the outlet grid cell in the network	tif	m

Elevation difference	Downstream stream node grid cell	outlet_diff_dw_scatch	Elevation difference between focal grid cell and the downstream stream node grid cell	tif	m

Segment properties	Segment downstream mean gradient	channel_grad_dw_seg	Segment downstream mean gradient (between focal cell and the node/outlet)	tif

Segment properties	Segment upstream mean gradient	channel_grad_up_seg	Segment upstream mean gradient (between focal cell and the init/node)	tif

Segment properties	Cell upstream gradient	channel_grad_up_cel	Cell upstream gradient (between focal cell and next cell	tif

                                                                                                                                  | Segment properties   | Cell stream course curvature             | channel curv_cel                    | Cell stream course curvature (focal cell)                                                           | tif         |      |
                                                                                                                                  | Segment properties   | Segment downstream elevation difference  | channel_elv_dw_seg                  | Segment downstream elevation difference (between focal cell and the node/outlet)                    | tif         |      |
                                                                                                                                  | Segment properties   | Segment upstream elevation difference    | channel_elv_up_seg                  | Segment upstream elevation difference (between focal cell and the init/node)                        | tif         |      |
                                                                                                                                  | Segment properties   | Cell upstream elevation difference       | channel_elv_up_cel                  | Cell upstream elevation difference (between focal cell and next cell)                               | tif         |      |
                                                                                                                                  | Segment properties   | Cell downstream elevation difference     | channel_elv_dw_cel                  | Cell downstream elevation difference (between focal cell and next cell)                             | tif         |      |
                                                                                                                                  | Segment properties   | Segment downstream distance              | channel_dist_dw_seg                 | Segment downstream distance (between focal cell and the node/outlet                                 | tif         |      |
                                                                                                                                                                                                                                                                           | Segment properties   | Segment upstream distance                | channel_dist_up_seg                 | Segment upstream distance (between focal cell and the init/node)                                    | tif         |      |
                                                                                                                                                                                                                                                                           | Segment properties   | Cell upstream distance                   | channel_dist_up_cel                 | Cell upstream distance (between focal cell and next cell)                                           | tif         |      |
                                                                                                                                                                                                                                                                           | Stream order         | Strahler’s stream order                  | order_strahler                      | Strahler’s stream order                                                                             | tif         |      |
                                                                                                                                                                                                                                                                           | Stream order         | Shreve’s stream magnitude                | order_shreve                        | Shreve’s stream magnitude                                                                           | tif         |      |
                                                                                                                                                                                                                                                                           | Stream order         | Horton’s stream order                    | order_horton                        | Horton’s stream order                                                                               | tif         |      |
                                                                                                                                                                                                                                                                           | Stream order         | Hack’s stream order                      | order_hack                          | Hack’s stream order                                                                                 | tif         |      |
                                                                                                                                                                                                                                                                           | Stream order         | Topological dimension of streams         | order_topo                          | Topological dimension of streams                                                                    | tif         |      |
                                                                                                                                                                                                                                                                           | Stream order         | Strahler’s stream order                  | order_vect_point/order_vect_segment | Strahler’s stream order                                                                             | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream order         | Shreve’s stream magnitude                | order_vect_point/order_vect_segment | Shreve’s stream magnitude                                                                           | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream order         | Horton’s stream order                    | order_vect_point/order_vect_segment | Horton’s stream order                                                                               | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream order         | Hack’s stream order                      | order_vect_point/order_vect_segment | Hack’s stream order                                                                                 | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream order         | Topological dimension of streams         | order_vect_point/order_vect_segment | Topological dimension of streams                                                                    | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream reach         | Length of the stream reach               | order_vect_point/order_vect_segment | Length of the stream reach                                                                          | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream reach         | Straight length                          | order_vect_point/order_vect_segment | Length of stream as straight line                                                                   | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream reach         | Sinusoid of the stream reach             | order_vect_point/order_vect_segment | Fractal dimension: stream length/straight stream length                                             | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream reach         | Accumulated length                       | order_vect_point/order_vect_segment | Length of stream from source                                                                        | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream reach         | Flow accumulation (?)                    | order_vect_point/order_vect_segment | Flow accumulation (what is the difference to the one above???)                                      | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream reach         | Distance to outlet                       | order_vect_point/order_vect_segment | Distance of current stream init from outlet                                                         | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream reach         | Source elevation                         | order_vect_point/order_vect_segment | Elevation of stream init                                                                            | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream reach         | Outlet elevation                         | order_vect_point/order_vect_segment | Elevation of stream outlet                                                                          | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream reach         | Elevation drop                           | order_vect_point/order_vect_segment | Difference between source_elev and outlet_elev + drop outlet                                        | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream reach         | Outlet drop                              | order_vect_point/order_vect_segment | Drop at the outlet of the stream                                                                    | gpkg        |      |
                                                                                                                                                                                                                                                                           | Stream reach         | Gradient                                 | order_vect_point/order_vect_segment | Mean gradient of the sub-catchment (downstream elevation difference divided by distance)            | gpkg        |      |
                                                                                                                                                                                                                                                                           | Flow index           | Stream power index                       | spi                                 | A measure of the erosive power of flowing water (Moore et al., 1991)                                | tif         |      |
                                                                                                                                                                                                                                                                           | Flow index           | Sediment transportation index            | sti                                 | A metric for describing the erosion and deposition of sediments (Mojaddadi et al., 2017)            | tif         |      |
                                                                                                                                                                                                                                                                           | Flow index           | Compound topographic index               | cti                                 | "A steady state wetness index, also known as topographic wetness index (TWI) (Beven & Kirkby,       |             |      |

References

Amatulli G., Garcia Marquez J., Sethi T., Kiesel J., Grigoropoulou A., Üblacker M., Shen L. & Domisch S. (2022-08-09 ) Hydrography90m: A new high-resolution global hydrographic dataset. IGB Leibniz-Institute of Freshwater Ecology and Inland Fisheries. dataset. https://doi.org/10.18728/igb-fred-762.1

Author

Afroditi Grigoropoulou

Examples

# Download data for two variables in three regular tiles
# to the current working directory
download_tiles(variable = c("sti", "stream_dist_up_farth"),
               file_format = "tif",
               tile_id = c("h00v02","h16v02", "h16v04"))

# Download the global .tif layer for the variable "direction"
# into the temporary R folder or define a different directory
# Define directory
my_directory <- tempdir()
# Download layer
download_tiles(variable = "direction",
               file_format = "tif",
               global = TRUE,
               download_dir = my_directory)