Extract lake ids

Extract lake ids either at the location of point occurrences or falling within a bounding box

Usage

extract_lake_ids(
  data,
  lon,
  lat,
  lake_shape,
  var_name = "Hylak_id",
  xmin = NULL,
  ymin = NULL,
  xmax = NULL,
  ymax = NULL,
  bbox = FALSE,
  lake_id_table,
  quiet = TRUE
)

Arguments

data

a data.frame or data.table that contains the columns the longitude / latitude coordinates in WGS84, i.e. of species point occurrences

lon

character. The name of the column with the longitude coordinates.

lat

character. The name of the column with the latitude coordinates.

lake_shape

Full path to lake shape files; i.e. HydroLAKES shapefiles; when NULL a lake raster file needs to be provided

var_name

character. Name of the shapefile ID column, i.e. "Hylak_id" for HydroLAKES; default is "Hylak_id"

xmin

integer. bounding box coordinate check what is lower left corner in WGS84; if NULL see parameter bbox

ymin

integer. bounding box coordinate check what is lower right corner in WGS84; if NULL see parameter bbox

xmax

integer. bounding box coordinate upper right corner in WGS84; if NULL see parameter bbox

ymax

integer. bounding box coordinate upper left corner in WGS84; if NULL see parameter bbox

bbox

logical. if TRUE, then all lake ids within the bounding box are returned, note that either bounding box values can be supplied thorugh xmin, ymin, xmax and ymax values or when left empty (NULL) bounding box is calculated from min max values of the longitude and latitude values provided in the occurrence table; if FALSE lake ids in which point occurrences are located are returned, default is FALSE

lake_id_table

character. Full path of the output lake id table

quiet

logical. If FALSE, the standard output will be printed. Default is TRUE.

lake_raster

Full path to lake raster file; note this option only works with bbox = FALSE and will result in the raster extraction at point occurrence locations; when NULL lake ids are extracted from lake shape file; default is NULL

read

logical. If TRUE, then the model .csv table gets read into R as data.table and data.frame. if FALSE, the table is only stored on disk. Default is FALSE.

References

add reference manual html here https://grass.osgeo.org/grass82/manuals/

Author

Jaime Garcia Marquez, Thomas Tomiczek

Examples