Calculate upstream basin

Calculates the upstream basin of a given point, considering the point as the outlet.

Usage

get_upstream_catchment(
  data,
  id,
  lon,
  lat,
  direction_layer = NULL,
  out_dir = NULL,
  n_cores = NULL,
  quiet = TRUE
)

Arguments

data

a data.frame or data.table with lat/lon coordinates in WGS84, which have been snapped to the stream network. The snapping can be done using the function 'snap_to_network'.

id

character. The name of a column containing unique IDs for each row of "data" (e.g., occurrence or site IDs).

lon

character. The name of the column with the longitude coordinates.

lat

character. The name of the column with the latitude coordinates.

direction_layer

character. Full path to raster file with the direction variable.

out_dir

Full path to the directory where the output(s) will be stored. To identify the upstream catchment the output file name includes the site id.

n_cores

numeric. Number of cores used for parallelization. If NULL, available cores - 1 will be used.

quiet

logical. If FALSE, the standard output will be printed. Default is TRUE.

References

https://grass.osgeo.org/grass82/manuals/r.water.outlet.html https://grass.osgeo.org/grass82/manuals/r.region.html

See also

Author

Jaime Garcia Marquez, Afroditi Grigoropoulou, Maria M. Üblacker

Examples

# Download test data into temporary R folder
# or define a different directory
my_directory <- tempdir()
download_test_data(my_directory)

# Before running the function get_upstream_catchment(), snap the points to
# to the stream segment. There are multiple ways to snap the points. Here is
# one example:

# Load occurrence data
species_occurence <- read.table(paste0(my_directory,
                                       "/hydrography90m_test_data",
                                       "/spdata_1264942.txt"),
                              header = TRUE)

# Define full path to the basin and sub-catchments raster layer
basin_raster <- paste0(my_directory,
                       "/hydrography90m_test_data/basin_1264942.tif")
subc_raster <- paste0(my_directory,
                      "/hydrography90m_test_data/subcatchment_1264942.tif")

# Define full path to the vector file of the stream network
stream_vector <- paste0(my_directory,
                        "/hydrography90m_test_data/order_vect_59.gpkg")

# Automatically extract the basin and sub-catchment IDs and
# snap the data points to the stream segment
snapped_coordinates <- snap_to_subc_segment(data = species_occurence,
                                            lon = "longitude",
                                            lat = "latitude",
                                            id = "occurrence_id",
                                            basin_layer = basin_raster,
                                            subc_layer = subc_raster,
                                            stream_layer = stream_vector,
                                            n_cores = 2)

# Define full path to the direction .tif
direction_raster <- paste0(my_directory,
                           "/hydrography90m_test_data/direction_1264942.tif")
# Define the path for the output file(s)
output_folder <-  paste0(my_directory, "/upstream_catchments")
if(!dir.exists(output_folder)) dir.create(output_folder)
# Get the upstream catchment for each point location
get_upstream_catchment(snapped_coordinates,
                       lon = "lon_snap",
                       lat = "lat_snap",
                       id = "occurrence_id",
                       direction_layer = direction_raster,
                       out_dir = output_folder,
                       n_cores = 2)